requirements.txt
pandas==2.1.1
numpy==1.26.3
pytorch==2.1.0
pyro-api==0.1.2
pyro-ppl==1.8.6
scipy==1.12.0
scikit-learn==1.3.0
pyyaml==6.0.1
tqdm==4.66.1
dataclasses==0.8